+++ source/Delphi.Forms.Info.pas
@@ -3,7 +3,7 @@ unit Delphi.Forms.Info;
 interface
 
 const
-  Version = '0.7.40.0';
+  Version = '0.7.42.0';
 
 implementation